pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "the-market-maker"
version = "0.1.0"
description = "A failure-aware, research-driven paper trading system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Market Maker Team"}
]
keywords = ["trading", "algorithmic", "paper-trading", "quantitative"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Data & APIs
    "alpaca-py>=0.21.0",
    "yfinance>=0.2.36",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    
    # Storage
    "duckdb>=0.10.0",
    "redis>=5.0.0",
    
    # Sentiment & NLP
    "praw>=7.7.0",           # Reddit API
    "tweepy>=4.14.0",        # Twitter API
    "transformers>=4.36.0",  # FinBERT
    "torch>=2.1.0",
    
    # Statistics & ML
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    
    # Utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
    "schedule>=1.2.0",
    
    # HTTP & Async
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.0",
    "mypy>=1.7.0",
    "ruff>=0.1.8",
    "black>=23.12.0",
    "pre-commit>=3.6.0",
]
research = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

[project.scripts]
market-maker = "src.main:main"
watchdog = "watchdog.daemon:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "watchdog*", "research*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
